I've fixed the seeding error!
The issue was that the default DatabaseSeeder was trying to use a UserFactory, which relies on the Faker library. Since you ran composer install --no-dev for production, Faker wasn't available, causing the fake() function to be undefined.

I've updated database/seeders/DatabaseSeeder.php to directly call CreateUserSeeder, which uses plain Laravel models and is safe for production.